+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @nrzt/cache
 
+## 2.1.19
+
+### Patch Changes
+
+- 将 qs 替换成 query-string
+- Updated dependencies
+  - @nrzt/cipher@2.1.8
+  - @nrzt/core@2.2.17
+
+## 2.1.18
+
+### Patch Changes
+
+- 修改依赖，提取出涉及的 dom 操作
+- Updated dependencies
+  - @nrzt/core@2.2.16
+  - @nrzt/cipher@2.1.7
+
 ## 2.1.17
 
 ### Patch Changes