+++ packages/request/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @nrzt/request
 
+## 2.1.20
+
+### Patch Changes
+
+- 将 qs 替换成 query-string
+- Updated dependencies
+  - @nrzt/core@2.2.17
+
+## 2.1.19
+
+### Patch Changes
+
+- 修改依赖，提取出涉及的 dom 操作
+- Updated dependencies
+  - @nrzt/core@2.2.16
+
 ## 2.1.18
 
 ### Patch Changes